pyproject.toml
[tool.poetry]
name = "trainer-back"
version = "0.1.0"
description = ""
authors = ["tomoki33"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
django = "^5.1.2"
mysqlclient = "^2.2.5"
djangorestframework = "^3.15.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
